Integrated code with zombiepaladin's commit
Added support for regular compressed .schematic files (and removed the uncompressed option)

Fixed schematic import bug

+++ Editor/Datatypes/VoxelTerrain.cs
@@ -42,9 +42,9 @@ namespace AweEditor
 
     public struct BlockData
     {
-        public short x, y, z;
+        public int x, y, z;
 
-        public short type;
+        public byte type;
     }
 
     /// <summary>
